requirements.txt
psycopg2-binary
python-dateutil
pandas
matplotlib
seaborn
plotly
reportlab
pytest